More refactoring
- What I've done -
Replaced the input boxes with something a little more effective and bug-free.
Added a simple timing system to the test cases in RandomWord and WordCheck.
Some optimizations made with generating the input boxes and a bug fix pertinent to the generation of the Wordle.
Documented more stuff and refactored existing documentation.

- Agenda -
There's a bug floating around causing a NullPointerException with the array functions in RandomWord and WordCheck. Not sure what's causing it, but in the test cases it produces a null array and throws the aforementioned exception. Could be read timeouts, I dunno.