Slight refactoring
- What I've Done -
Made some more optimizations, added the ability to replace a character in a character box without having to explicitly and manually delete its contents, and tidied up some documentation.

- Agenda -
The "bug" returning a null array in the test cases was likely due to network issues. I'll look into using the Future class or the like to try and more reliably solve this. Other than that, I plan on setting the cursor focus onto the next character box after the player has typed in a character - just for convenience sake.