Refactoring and serialization removal
- What I've Done -
Implemented Future class and asynchronous function calling using ExecutorService to build a simple thread with a lambda attached. This is to catch the bug mentioned in the previous commit.
Removed JFrame's inheritance into Wordle; in other words, no more "Wordle extends JFrame," thus removing the annoying serialization warnings.

- Agenda -
Rethink the whole implementation of the Future class. Looks a little ugly to me when a try-catch could probably achieve the same result at the cost of maybe a few milliseconds.